Obtain BD data from a new source in GitHub
The versioning also changed and we now use Long. The version represents the date and time of BD generation.

+++ SeviBus/src/main/java/com/sloy/sevibus/ui/activities/ContactoActivity.java
@@ -116,7 +116,7 @@ public class ContactoActivity extends BaseToolbarActivity {
         StringBuilder texto = new StringBuilder(mMensaje.getText().toString().trim());
         texto.append("\n\n=====================");
         texto.append("\n Versión: " + BuildConfig.VERSION_NAME);
-        texto.append("\n Base de datos: " + getSharedPreferences("datos", Context.MODE_MULTI_PROCESS).getInt("data_version", 0));
+        texto.append("\n Base de datos: " + getSharedPreferences("datos", Context.MODE_MULTI_PROCESS).getLong("data_version", 0));
         if (mDeviceInfo.isChecked()) {
             texto.append("\n\nInformación del dispositivo");
             texto.append("\n---------- ");
